backend/requirements/common.txt
django~=3.0.0
djangorestframework
djangorestframework-jwt
django-rest-authtoken
django-cors-headers
pillow